Rework project structure

+++ 8_clean_arch/internal/models/special.go
@@ -1,8 +1,8 @@
 package models
 
 type Special struct {
-	Id        int    `json:"id"`
-	ProductId int    `json:"product_id"`
+	Id        uint64 `json:"id"`
+	ProductId uint64 `json:"product_id"`
 	Slogan    string `json:"slogan"`
 	Discount  int    `json:"discount"`
 }
